requirements.txt
streamlit
python-dotenv
groq
PyPDF2
python-docx
python-pptx